Back to back key presses are now honored
Previously, if two keys were pressed back to back within the
movement window, e.g. Left and then Up, only the second key
was taken for the movement.

This is incorrect as it results in a key press loss.

Now the key press results are queued so that the player can do fast
key presses.